spring-start/hello-spring/src/test/java/hello/hellospring/Repository/MemoryMemberRepositoryTest.java
package hello.hellospring.Repository;

import hello.hellospring.domain.Member;
import hello.hellospring.repository.MemoryMemberRepository;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Test;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;

public class MemoryMemberRepositoryTest {
    MemoryMemberRepository repository = new MemoryMemberRepository();

    @AfterEach
    public void afterEach() {
        repository.clearStore();
    }

    @Test
    public void save() {
        Member member = new Member();
        member.setName("spring");//멤버의 이름을 spring으로 설정

        repository.save(member);
        Member result = repository.findById(member.getId()).get();
//        System.out.println("result = " + (result == member));
        assertThat(member).isEqualTo(result);

    }

    @Test
    public void findByName() {
        Member member = new Member();
        member.setName("Spring1");
        repository.save(member);

        Member member2 = new Member();
        member2.setName("Spring2");
        repository.save(member2);

        Member result = repository.findByName("Spring1").get();
        //assertThat(spring1).isEqualTo(member);
        System.out.println("spring1 = " + (result == member));

    }

    @Test
    public void findAll() {
        Member member = new Member();
        member.setName("Spring3");
        repository.save(member);

        Member member2 = new Member();
        member2.setName("Spring4");
        repository.save(member2);

        List<Member> result = repository.findAll();
        assertThat(result.size()).isEqualTo(2);
    }
}